+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Added
 
 - [TD-3945] New permissions to view and manage grants
+- [TD-3947] Updated td-cache to write cache for user_name
 
 ## [4.24.0] 2021-07-13